started over to add better error handling
the last version did a very poor job of handling errors, especially when using
maps. Instead of throwing an error if a key is not found in a map, golang
returns two values. The first value is map[key] and the second (optional
argument) is true if the key was found. If the key was not found, the first
value is the value type's zero value. I've gone through the code generation
functions and added checks using the second argument before using the value
returned by the maps. I still need to do this for pushPop(). Also added a
data directive to the assembler to allow raw data to be inserted easily. I may
also extend this to different types of directives.