started making demo.asm into pong, added directives for variables and aliases
variables, savedVariables and aliases are all just fancy names for labels.
aliases can be labels to any address, but they don't run the risk of messing
up the program by changing the assembler's program counter. variables are labels
pointing to addresses within work ram (8 kb starting at 0xc000). addresses are
automatically assigned by the assembler and each variable can occupy 1 or 2 bytes.
savedVariables are the same thing as variables but with an address in external ram
(8 kB starting at 0xa000). the difference between this and work ram is that external
ram is saved each time the game is closed and reloaded when it's started. also added
cpl instruction which is very useful for negating numbers and shorter than xor 0xff