created allocateVariable() for .var and .save directives
also decided to make all variables labels in the top scope level for now. Adding
scopes for variables is straightforward since they are just labels to various
RAM banks but I need to make sure to update wram/eram_counter as needed at the
end of each scope. This may be useful for writing directives for higher level
abstractions like for/while loops but I need to think about whether this is
worth the effort